backend/server/pyproject.toml
[project]
name = "songbird-backend"
version = "0.1.0"
description = "SongBird backend server"
requires-python = ">=3.12, <4"
dependencies = [
    # Web
    "fastapi>=0.109.1",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.3.0",
    "slowapi>=0.1.9",
    # Config
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    # Database
    "SQLAlchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "GeoAlchemy2>=0.15.0",
    "shapely>=2.0.0",
    "greenlet>=2.0.2",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.1",
    # Redis
    "redis>=5.0.1",
    # HTTP client (external APIs)
    "httpx>=0.26.0",
    # Notifications
    "twilio>=9.0.0",
    # Logging
    "structlog>=25.1.0",
    "rich>=14.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
    "faker>=26.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "C"]
ignore = ["E402", "UP006", "UP007", "E741"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
